hollow/My project/whisper_server/requirements.txt
flask>=3.0.0
openai-whisper>=20231117
